backend/pyproject.toml
[tool.poetry]
name = "Recipe"
version = "0.1.0"
description = ""
authors = [""]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
SQLAlchemy = "^1.4.25"
uvicorn = "^0.15.0"
asyncpg = "^0.24.0"
alembic = "^1.7.4"
pydantic = "^1.8.2"
psycopg2-binary = "^2.9.2"
python-dotenv = "^0.19.2"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
social-auth-core = { extras = ["openidconnect"], version = "^4.1.0" }
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
httpx = "^0.19.0"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
